Refactor: Color Recommend Screen 리팩토링
