+++ src/utils/ColorRecommendScreen/infoText.jsx
@@ -1,0 +1,2 @@
+export const infoText =
+	'• 선택한 색상을 기반하여 색상 추천 목록을 제공합니다.\n• 원하는 색상을 터치해 색상의 자세한 정보를 확인하세요!';
